Using controllers to organize user routes
